+++ docs/intezer_sdk.rst
@@ -7,7 +7,7 @@ Submodules
 khulnasoft\_sdk.account module
 -------------------------------------
 
-.. automodule:: khulnasoft_sdk.account
+.. automodule:: khulnasoft_analyze_sdk.account
    :members:
    :undoc-members:
    :show-inheritance:
@@ -15,7 +15,7 @@ khulnasoft\_sdk.account module
 khulnasoft\_sdk.analyses\_history module
 -------------------------------------
 
-.. automodule:: khulnasoft_sdk.analyses_history
+.. automodule:: khulnasoft_analyze_sdk.analyses_history
    :members:
    :undoc-members:
    :show-inheritance:
@@ -24,7 +24,7 @@ khulnasoft\_sdk.analyses\_history module
 khulnasoft\_sdk.analyses\_results module
 -------------------------------------
 
-.. automodule:: khulnasoft_sdk.analyses_results
+.. automodule:: khulnasoft_analyze_sdk.analyses_results
    :members:
    :undoc-members:
    :show-inheritance:
@@ -32,7 +32,7 @@ khulnasoft\_sdk.analyses\_results module
 khulnasoft\_sdk.analysis module
 ----------------------------
 
-.. automodule:: khulnasoft_sdk.analysis
+.. automodule:: khulnasoft_analyze_sdk.analysis
    :members:
    :undoc-members:
    :show-inheritance:
@@ -41,7 +41,7 @@ khulnasoft\_sdk.analysis module
 khulnasoft\_sdk.api module
 -----------------------
 
-.. automodule:: khulnasoft_sdk.api
+.. automodule:: khulnasoft_analyze_sdk.api
    :members:
    :undoc-members:
    :show-inheritance:
@@ -50,7 +50,7 @@ khulnasoft\_sdk.api module
 khulnasoft\_sdk.base\_analysis module
 ----------------------------------
 
-.. automodule:: khulnasoft_sdk.base_analysis
+.. automodule:: khulnasoft_analyze_sdk.base_analysis
    :members:
    :undoc-members:
    :show-inheritance:
@@ -58,7 +58,7 @@ khulnasoft\_sdk.base\_analysis module
 khulnasoft\_sdk.consts module
 --------------------------
 
-.. automodule:: khulnasoft_sdk.consts
+.. automodule:: khulnasoft_analyze_sdk.consts
    :members:
    :undoc-members:
    :show-inheritance:
@@ -66,7 +66,7 @@ khulnasoft\_sdk.consts module
 khulnasoft\_sdk.endpoint\_analysis module
 --------------------------------------
 
-.. automodule:: khulnasoft_sdk.endpoint_analysis
+.. automodule:: khulnasoft_analyze_sdk.endpoint_analysis
    :members:
    :undoc-members:
    :show-inheritance:
@@ -74,7 +74,7 @@ khulnasoft\_sdk.endpoint\_analysis module
 khulnasoft\_sdk.errors module
 --------------------------
 
-.. automodule:: khulnasoft_sdk.errors
+.. automodule:: khulnasoft_analyze_sdk.errors
    :members:
    :undoc-members:
    :show-inheritance:
@@ -82,7 +82,7 @@ khulnasoft\_sdk.errors module
 khulnasoft\_sdk.family module
 --------------------------
 
-.. automodule:: khulnasoft_sdk.family
+.. automodule:: khulnasoft_analyze_sdk.family
    :members:
    :undoc-members:
    :show-inheritance:
@@ -90,7 +90,7 @@ khulnasoft\_sdk.family module
 khulnasoft\_sdk.index module
 -------------------------
 
-.. automodule:: khulnasoft_sdk.index
+.. automodule:: khulnasoft_analyze_sdk.index
    :members:
    :undoc-members:
    :show-inheritance:
@@ -98,7 +98,7 @@ khulnasoft\_sdk.index module
 khulnasoft\_sdk.operation module
 -----------------------------
 
-.. automodule:: khulnasoft_sdk.operation
+.. automodule:: khulnasoft_analyze_sdk.operation
    :members:
    :undoc-members:
    :show-inheritance:
@@ -106,7 +106,7 @@ khulnasoft\_sdk.operation module
 khulnasoft\_sdk.sub\_analysis module
 ---------------------------------
 
-.. automodule:: khulnasoft_sdk.sub_analysis
+.. automodule:: khulnasoft_analyze_sdk.sub_analysis
    :members:
    :undoc-members:
    :show-inheritance:
@@ -114,7 +114,7 @@ khulnasoft\_sdk.sub\_analysis module
 khulnasoft\_sdk.util module
 ------------------------
 
-.. automodule:: khulnasoft_sdk.util
+.. automodule:: khulnasoft_analyze_sdk.util
    :members:
    :undoc-members:
    :show-inheritance:
@@ -122,7 +122,7 @@ khulnasoft\_sdk.util module
 Module contents
 ---------------
 
-.. automodule:: khulnasoft_sdk
+.. automodule:: khulnasoft_analyze_sdk
    :members:
    :undoc-members:
    :show-inheritance:
