+++ docs/modules.rst
@@ -1,7 +1,7 @@
-khulnasoft_sdk
+khulnasoft_analyze_sdk
 ===========
 
 .. toctree::
    :maxdepth: 4
 
-   khulnasoft_sdk
+   khulnasoft_analyze_sdk
