+++ khulnasoft_analyze_sdk/_account_api.py
@@ -2,10 +2,10 @@ from http import HTTPStatus
 from typing import List
 from typing import Optional
 
-from khulnasoft_sdk import errors
-from khulnasoft_sdk.api import KhulnasoftApiClient
-from khulnasoft_sdk.api import get_global_api
-from khulnasoft_sdk.api import raise_for_status
+from khulnasoft_analyze_sdk import errors
+from khulnasoft_analyze_sdk.api import KhulnasoftApiClient
+from khulnasoft_analyze_sdk.api import get_global_api
+from khulnasoft_analyze_sdk.api import raise_for_status
 
 
 class AccountApi:
