+++ khulnasoft_analyze_sdk/_api.py
@@ -11,13 +11,13 @@ from typing import Tuple
 
 from requests import Response
 
-from khulnasoft_sdk import consts
-from khulnasoft_sdk import errors
-from khulnasoft_sdk._util import deprecated
-from khulnasoft_sdk.api import KhulnasoftApiClient
-from khulnasoft_sdk.api import raise_for_status
-from khulnasoft_sdk.consts import IndexType
-from khulnasoft_sdk.consts import OnPremiseVersion
+from khulnasoft_analyze_sdk import consts
+from khulnasoft_analyze_sdk import errors
+from khulnasoft_analyze_sdk._util import deprecated
+from khulnasoft_analyze_sdk.api import KhulnasoftApiClient
+from khulnasoft_analyze_sdk.api import raise_for_status
+from khulnasoft_analyze_sdk.consts import IndexType
+from khulnasoft_analyze_sdk.consts import OnPremiseVersion
 
 
 class KhulnasoftApi:
