+++ tests/unit/test_index.py
@@ -5,9 +5,9 @@ from unittest.mock import patch
 
 import responses
 
-from khulnasoft_sdk import consts
-from khulnasoft_sdk import errors
-from khulnasoft_sdk.index import Index
+from khulnasoft_analyze_sdk import consts
+from khulnasoft_analyze_sdk import errors
+from khulnasoft_analyze_sdk.index import Index
 from tests.unit.base_test import BaseTest
