+++ tests/unit/test_offline_endpoint_analysis.py
@@ -4,9 +4,9 @@ from http import HTTPStatus
 
 import responses
 
-from khulnasoft_sdk import consts
-from khulnasoft_sdk.endpoint_analysis import EndpointAnalysis
-from khulnasoft_sdk.errors import KhulnasoftError
+from khulnasoft_analyze_sdk import consts
+from khulnasoft_analyze_sdk.endpoint_analysis import EndpointAnalysis
+from khulnasoft_analyze_sdk.errors import KhulnasoftError
 from tests.unit.base_test import BaseTest
