+++ tests/unit/test_results.py
@@ -7,17 +7,17 @@ from typing import List
 
 import responses
 
-from khulnasoft_sdk.alerts import query_alerts_history
-from khulnasoft_sdk.alerts import ALERTS_SEARCH_REQUEST
-from khulnasoft_sdk.analyses_history import ENDPOINT_ANALYSES_REQUEST
-from khulnasoft_sdk.analyses_history import URL_ANALYSES_REQUEST
-from khulnasoft_sdk.analyses_history import generate_analyses_history_filter
-from khulnasoft_sdk.analyses_history import FILE_ANALYSES_REQUEST
-from khulnasoft_sdk.analyses_history import query_endpoint_analyses_history
-from khulnasoft_sdk.analyses_history import query_file_analyses_history
-from khulnasoft_sdk.analyses_history import query_url_analyses_history
-from khulnasoft_sdk.api import get_global_api
-from khulnasoft_sdk.analyses_results import AnalysesHistoryResult
+from khulnasoft_analyze_sdk.alerts import query_alerts_history
+from khulnasoft_analyze_sdk.alerts import ALERTS_SEARCH_REQUEST
+from khulnasoft_analyze_sdk.analyses_history import ENDPOINT_ANALYSES_REQUEST
+from khulnasoft_analyze_sdk.analyses_history import URL_ANALYSES_REQUEST
+from khulnasoft_analyze_sdk.analyses_history import generate_analyses_history_filter
+from khulnasoft_analyze_sdk.analyses_history import FILE_ANALYSES_REQUEST
+from khulnasoft_analyze_sdk.analyses_history import query_endpoint_analyses_history
+from khulnasoft_analyze_sdk.analyses_history import query_file_analyses_history
+from khulnasoft_analyze_sdk.analyses_history import query_url_analyses_history
+from khulnasoft_analyze_sdk.api import get_global_api
+from khulnasoft_analyze_sdk.analyses_results import AnalysesHistoryResult
 from tests.unit.base_test import BaseTest
