+++ tests/unit/test_url_analysis.py
@@ -3,11 +3,11 @@ from http import HTTPStatus
 
 import responses
 
-from khulnasoft_sdk import consts
-from khulnasoft_sdk import errors
-from khulnasoft_sdk.analysis import UrlAnalysis
-from khulnasoft_sdk.api import get_global_api
-from khulnasoft_sdk.consts import OnPremiseVersion
+from khulnasoft_analyze_sdk import consts
+from khulnasoft_analyze_sdk import errors
+from khulnasoft_analyze_sdk.analysis import UrlAnalysis
+from khulnasoft_analyze_sdk.api import get_global_api
+from khulnasoft_analyze_sdk.consts import OnPremiseVersion
 from tests.unit.base_test import BaseTest
